requirements-testing.txt
codecov
pytest-django
pytest-cov
dj-database-url==0.4.2
gunicorn==19.7.1
psycopg2==2.7.3.2
whitenoise==3.3.1
django-storages==1.6.5
boto3==1.5.1
Pillow==4.3.0
Markdown==2.6.10
cloudinary==1.9.0
requests==2.18.4
graphene-django==2.0.0
graphene==2.0.1
singlemodeladmin==0.8
boto
feedparser==5.2.1
authy==2.1.5
django-phonenumber-field==1.3.0
django-taggit==0.22.2
pytz==2017.3